+++ src/app/pages/tabs/analysis/analysis-form/analysis-form.page.ts
@@ -85,6 +85,7 @@ export class AnalysisFormPage {
    */
   ionViewWillLeave(): void {
     this.sbs.forEach((s) => s.unsubscribe());
+    this.updateReportEvent.complete();
   }
 
   /**