+++ src/app/pages/tabs/analysis/take-photos/take-photos.page.ts
@@ -1,5 +1,5 @@
-import { Component, OnDestroy, OnInit } from "@angular/core";
-import { Camera, CameraResultType, CameraSource } from "@capacitor/camera";
+import { Component } from "@angular/core";
+import { CameraSource } from "@capacitor/camera";
 import { UtilsService } from "src/app/services/utils.service";
 import { FormControl, Validators } from "@angular/forms";
 import { ReportService } from "src/app/services/report.service";
@@ -18,7 +18,7 @@ import { CameraService } from "src/app/services/camera.service";
   templateUrl: "./take-photos.page.html",
   styleUrls: ["./take-photos.page.scss"],
 })
-export class TakePhotosPage implements OnInit, OnDestroy {
+export class TakePhotosPage {
   especie = new FormControl("", [Validators.required]);
   photo = new FormControl("");
   loadingPhoto: boolean;
@@ -53,18 +53,27 @@ export class TakePhotosPage implements OnInit, OnDestroy {
     this.hasActiveWood$ = this.reportService.activeWood.pipe(
       map((wood) => !!wood)
     );
-    this.saveWoodHandler();
   }
 
-  ngOnInit() {
+  /**
+   * Build subscriptions/event handlers for component, every time Page is 'Entered'.
+   */
+  ionViewWillEnter(): void {
+    this.saveWoodHandler();
     this.populateForm();
   }
 
-  ngOnDestroy(): void {
+  /**
+   * Destroy subscriptions/event handlers for component, every time Page is 'Left'.
+   */
+  ionViewWillLeave(): void {
     this.sbs.forEach((s) => s.unsubscribe());
     this.saveWoodEvent.complete();
   }
 
+  /**
+   * Populate form with active wood data.
+   */
   populateForm() {
     this.sbs.push(
       this.reportService.activeWood