+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # [RED-23] -2023-07-13 (Mario)
 
+- AnalysisPage/AnalysisResultPage/TakePhotosPage: Manejar subscripciones con el ciclo de vida de ionic.
 - AnalysisFormPage: Fix para que se actualice el campo 'firstReport' en el WtUser, al crear el primer report.
 - capacitor.config.ts: Fix backgroundColor.